Replaced size_t with unsigned long here too.

+++ src/multi_type_vector_test_custom.cpp
@@ -330,7 +330,7 @@ void mtv_test_types()
     assert(ct == mtv::element_type_numeric);
     ct = column_type::get_element_type(string());
     assert(ct == mtv::element_type_string);
-    ct = column_type::get_element_type(size_t(12));
+    ct = column_type::get_element_type(static_cast<unsigned long>(12));
     assert(ct == mtv::element_type_ulong);
     ct = column_type::get_element_type(true);
     assert(ct == mtv::element_type_boolean);
@@ -579,7 +579,7 @@ void mtv_test_managed_block()
         db.set_empty(0, 1);
 
         db.set(0, new muser_cell(6.0));
-        db.set(1, size_t(12));
+        db.set(1, static_cast<unsigned long>(12));
         db.set_empty(0, 2);
     }
 
@@ -846,7 +846,7 @@ void mtv_test_managed_block()
         column_type db(4);
         db.set(0, 1.1);
         db.set(1, new muser_cell(1.0));
-        db.set(2, size_t(2));
+        db.set(2, static_cast<unsigned long>(2));
         db.set(3, new muser_cell(3.0));
         assert(db.block_size() == 4);
         assert(db.size() == 4);
@@ -867,15 +867,15 @@ void mtv_test_managed_block()
         db.set(2, 1.2);
         db.set(3, new muser_cell(3.0));
         db.set(4, new muser_cell(4.0));
-        size_t vals[] = { 5, 6, 7 };
-        const size_t* p = &vals[0];
+        unsigned long vals[] = { 5, 6, 7 };
+        const unsigned long* p = &vals[0];
         db.set(1, p, p+3);
     }
 
     {
         // set_cells() across multiple blocks, part 2.
         column_type db(6);
-        db.set(0, size_t(12));
+        db.set(0, static_cast<unsigned long>(12));
         db.set(1, new muser_cell(1.0));
         db.set(2, new muser_cell(2.0));
         db.set(3, 1.2);
@@ -894,14 +894,14 @@ void mtv_test_managed_block()
     {
         // set_cell() to merge 3 blocks.
         column_type db(6);
-        db.set(0, size_t(12));
+        db.set(0, static_cast<unsigned long>(12));
         db.set(1, new muser_cell(1.0));
         db.set(2, new muser_cell(2.0));
         db.set(3, 1.2);
         db.set(4, new muser_cell(3.0));
         db.set(5, new muser_cell(4.0));
         assert(db.block_size() == 4);
-        assert(db.get<size_t>(0) == 12);
+        assert(db.get<unsigned long>(0) == 12);
         assert(db.get<muser_cell*>(1)->value == 1.0);
         assert(db.get<muser_cell*>(2)->value == 2.0);
         assert(db.get<double>(3) == 1.2);
@@ -910,7 +910,7 @@ void mtv_test_managed_block()
 
         db.set(3, new muser_cell(5.0)); // merge blocks.
         assert(db.block_size() == 2);
-        assert(db.get<size_t>(0) == 12);
+        assert(db.get<unsigned long>(0) == 12);
         assert(db.get<muser_cell*>(1)->value == 1.0);
         assert(db.get<muser_cell*>(2)->value == 2.0);
         assert(db.get<muser_cell*>(3)->value == 5.0);
@@ -921,7 +921,7 @@ void mtv_test_managed_block()
     {
         // set_cell() to merge 2 blocks.
         column_type db(3);
-        db.set(0, size_t(23));
+        db.set(0, static_cast<unsigned long>(23));
         db.set(1, new muser_cell(2.1));
         db.set(2, new muser_cell(3.1));
